Remove hook_webform_paymethod_select_redirect_alter().
The hook was never documented because the *.api.php file was
accidentally not committed. The only usage of this hook has been
removed in favor of hook_webform_redirect_alter() provided by
little_helpers (>=2.0-alpha1). We already depend on little_helpers
(>=2.0-alpha3) so there is no need to pump the dependency.

Webform::getRedirect() always returns an array so the code can be
simplified even further.